requirements.txt
opencv-python==4.8.0.76
mediapipe==0.10.5
pyserial==3.5
protobuf==3.20.3